Added a util.c and a progress bar
