+++ src/include/main.h
@@ -6,7 +6,6 @@
 #define MB (1024 * KB)
 #define GB (1024 * MB)
 #define TB (1024 * GB)
-#define BUFFER_SIZE (1024 * 1024 * 32) // 32MB
 
 struct DISK_INFO {
     unsigned long long size;
